Move FML lang loader version methods into base class + spotlessApply
