+++ loader/src/main/java/net/neoforged/fml/loading/LanguageProviderLoader.java
@@ -6,14 +6,9 @@
 package net.neoforged.fml.loading;
 
 import com.mojang.logging.LogUtils;
-import java.net.URISyntaxException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.stream.Stream;
